# tests/golden/openapi_schema.json
{
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
      "MetricComparisonRow": {
        "properties": {
          "candidate": {
            "title": "Candidate",
            "type": "number"
          },
          "champion": {
            "title": "Champion",
            "type": "number"
          },
          "delta": {
            "title": "Delta",
            "type": "number"
          },
          "status": {
            "title": "Status",
            "type": "string"
          }
        },
        "required": [
          "champion",
          "candidate",
          "delta",
          "status"
        ],
        "title": "MetricComparisonRow",
        "type": "object"
      },
...
      "SimulationRequest": {
        "properties": {
          "override_thresholds": {
            "additionalProperties": {
              "type": "number"
            },
            "title": "Override Thresholds",
            "type": "object"
          },
          "run_id": {
            "title": "Run Id",
            "type": "string"
          }
        },
        "required": [
          "run_id"
        ],
        "title": "SimulationRequest",
        "type": "object"
      },
      "SimulationResponse": {
        "properties": {
          "champion_metrics": {
            "additionalProperties": {
              "type": "number"
            },
            "title": "Champion Metrics",
            "type": "object"
          },
          "config_snapshot": {
            "additionalProperties": true,
            "title": "Config Snapshot",
            "type": "object"
          },
          "decision": {
            "title": "Decision",
            "type": "string"
          },
          "decision_narrative": {
            "default": "",
            "title": "Decision Narrative",
            "type": "string"
          },
          "delta_metrics": {
            "additionalProperties": {
              "type": "number"
            },
            "title": "Delta Metrics",
            "type": "object"
          },
          "is_first_run": {
            "default": false,
            "title": "Is First Run",
            "type": "boolean"
          },
          "metric_table": {
            "additionalProperties": {
              "$ref": "#/components/schemas/MetricComparisonRow"
            },
            "title": "Metric Table",
            "type": "object"
          },
          "new_metrics": {
            "additionalProperties": {
              "type": "number"
            },
            "title": "New Metrics",
            "type": "object"
          },
          "reason": {
            "title": "Reason",
            "type": "string"
          },
          "run_id": {
            "title": "Run Id",
            "type": "string"
          },
          "should_promote": {
            "default": false,
            "title": "Should Promote",
            "type": "boolean"
          },
          "slice_regressions": {
            "items": {
              "additionalProperties": true,
              "type": "object"
            },
            "title": "Slice Regressions",
            "type": "array"
          }
        },
        "required": [
          "decision",
          "reason",
          "run_id"
        ],
        "title": "SimulationResponse",
        "type": "object"
      },
...
      "ThresholdHistoricalRunPayload": {
        "properties": {
          "candidate_metrics": {
            "additionalProperties": {
              "type": "number"
            },
            "title": "Candidate Metrics",
            "type": "object"
          },
          "champion_metrics": {
            "additionalProperties": {
              "type": "number"
            },
            "title": "Champion Metrics",
            "type": "object"
          },
          "should_pass": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Should Pass"
          },
          "slice_metrics": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Slice Metrics"
          }
        },
        "required": [
          "candidate_metrics",
          "champion_metrics"
        ],
        "title": "ThresholdHistoricalRunPayload",
        "type": "object"
      },
      "ThresholdOptimizeRequest": {
        "properties": {
          "accuracy_tolerances": {
            "anyOf": [
              {
                "items": {
                  "type": "number"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Accuracy Tolerances"
          },
          "critical_slice_floors": {
            "anyOf": [
              {
                "items": {
                  "type": "number"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Critical Slice Floors"
          },
          "f1_tolerances": {
            "anyOf": [
              {
                "items": {
                  "type": "number"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "F1 Tolerances"
          },
          "historical_runs": {
            "items": {
              "$ref": "#/components/schemas/ThresholdHistoricalRunPayload"
            },
            "minItems": 1,
            "title": "Historical Runs",
            "type": "array"
          }
        },
        "required": [
          "historical_runs"
        ],
        "title": "ThresholdOptimizeRequest",
        "type": "object"
      },
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      }
    }
  },
  "info": {
    "title": "Ares",
    "version": "1.0.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/gate/config": {
      "get": {
        "operationId": "get_gate_config_api_v1_gate_config_get",
        "parameters": [
          {
            "in": "header",
            "name": "X-API-Key",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Api-Key"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Get Gate Config Api V1 Gate Config Get",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Gate Config",
        "tags": [
          "gate"
        ]
      }
    },
    "/api/v1/gate/optimize": {
      "post": {
        "operationId": "optimize_gate_thresholds_api_v1_gate_optimize_post",
        "parameters": [
          {
            "in": "header",
            "name": "X-API-Key",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Api-Key"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ThresholdOptimizeRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Optimize Gate Thresholds Api V1 Gate Optimize Post",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Optimize Gate Thresholds",
        "tags": [
          "gate"
        ]
      }
    },
    "/api/v1/gate/simulate": {
      "post": {
        "operationId": "simulate_gate_api_v1_gate_simulate_post",
        "parameters": [
          {
            "in": "header",
            "name": "X-API-Key",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Api-Key"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SimulationRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SimulationResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Simulate Gate",
        "tags": [
          "gate"
        ]
      }
    },
...
  }
}